Move InconsistentLicensePoolState to palace.manager.core.exceptions (PP-3977)
palace-util is a generic shared package and should not carry LicensePool-domain
exceptions. Move the exception to src/palace/manager/core/exceptions.py where
other domain-specific exceptions like IntegrationException already live.
