+++ tests/manager/celery/tasks/test_apply.py
@@ -2,9 +2,10 @@ from unittest.mock import patch
 
 import pytest
 
-from palace.util.exceptions import InconsistentLicensePoolState, PalaceTypeError
+from palace.util.exceptions import PalaceTypeError
 
 from palace.manager.celery.tasks import apply
+from palace.manager.core.exceptions import InconsistentLicensePoolState
 from palace.manager.data_layer.bibliographic import BibliographicData
 from palace.manager.data_layer.circulation import CirculationData
 from palace.manager.data_layer.identifier import IdentifierData
